refactor(edge): isolate route resolution and backend selection
