+++ crates/edge/src/quic_listener/forwarding/mod.rs
@@ -3099,230 +3099,6 @@ impl QUICListener {
         Ok(())
     }
 
-    pub(super) fn resolve_backend_without_inflight(
-        method: &str,
-        path: &str,
-        authority: Option<&str>,
-        cid_key: Option<&str>,
-        upstream_pools: &HashMap<String, Arc<RwLock<UpstreamPool>>>,
-        routing_index: &RouteIndex,
-        header_lookup: Option<&LbHeaderLookup<'_>>,
-    ) -> Result<ResolvedBackend, ProxyError> {
-        if method.is_empty() || path.is_empty() {
-            return Err(ProxyError::Transport("empty method or path".into()));
-        }
-
-        let route_decision = routing_index
-            .lookup_with_decision_for_method(path, authority, Some(method))
-            .ok_or_else(|| ProxyError::Transport(format!("no route for {path}")))?;
-        let upstream_name = route_decision.upstream;
-
-        let upstream_pool = upstream_pools
-            .get(upstream_name)
-            .ok_or_else(|| ProxyError::Transport(format!("pool not found: {upstream_name}")))?
-            .clone();
-
-        let (backend_index, lb_type, backend_addr) = {
-            let (read_lb_type, read_fast_selected) = {
-                let pool = upstream_pool
-                    .read()
-                    .map_err(|_| ProxyError::Transport("upstream pool lock poisoned".into()))?;
-                if pool.pool.is_empty() {
-                    return Err(ProxyError::Transport("no servers in upstream".into()));
-                }
-                let lb_type = pool.lb_name();
-                let key = Self::resolve_lb_request_key(
-                    lb_type,
-                    pool.lb_key(),
-                    method,
-                    path,
-                    authority,
-                    cid_key,
-                    header_lookup,
-                );
-                let fast_selected = if pool.pool.readmit_due() {
-                    None
-                } else {
-                    pool.pick_readonly(key.as_str())
-                        .and_then(|idx| pool.pool.address(idx).map(|addr| (idx, addr.to_string())))
-                };
-                (lb_type, fast_selected)
-            };
-
-            if let Some((idx, addr)) = read_fast_selected {
-                (idx, read_lb_type, addr)
-            } else {
-                let mut pool = upstream_pool
-                    .write()
-                    .map_err(|_| ProxyError::Transport("upstream pool lock poisoned".into()))?;
-                if pool.pool.is_empty() {
-                    return Err(ProxyError::Transport("no servers in upstream".into()));
-                }
-                let lb_type = pool.lb_name();
-                let key = Self::resolve_lb_request_key(
-                    lb_type,
-                    pool.lb_key(),
-                    method,
-                    path,
-                    authority,
-                    cid_key,
-                    header_lookup,
-                );
-                let idx = pool.pick_without_begin(key.as_str()).ok_or_else(|| {
-                    let total = pool.pool.len();
-                    let healthy = pool.pool.healthy_len();
-                    error!(
-                        "no healthy backends available: {}/{} backends healthy",
-                        healthy, total
-                    );
-                    ProxyError::Transport("no healthy servers".into())
-                })?;
-                let backend_addr = pool
-                    .pool
-                    .address(idx)
-                    .map(str::to_string)
-                    .ok_or_else(|| ProxyError::Transport("invalid server address".into()))?;
-                (idx, lb_type, backend_addr)
-            }
-        };
-
-        debug!(
-            "Selected backend {} via {} route={} path_len={} host_specific={} reason={:?}",
-            backend_addr,
-            lb_type,
-            upstream_name,
-            route_decision.matched_path_len,
-            route_decision.host_specific,
-            route_decision.reason
-        );
-        Ok(ResolvedBackend {
-            upstream_name: upstream_name.to_string(),
-            backend_addr,
-            backend_index,
-            upstream_pool,
-            backend_lb: lb_type.to_string(),
-            route_path_len: route_decision.matched_path_len,
-            route_host_specific: route_decision.host_specific,
-            route_reason: route_decision.reason,
-        })
-    }
-
-    /// Resolve routing + LB for a request, returning `(backend_addr, backend_index, pool)`.
-    pub(super) fn resolve_backend(
-        method: &str,
-        path: &str,
-        authority: Option<&str>,
-        cid_key: Option<&str>,
-        upstream_pools: &HashMap<String, Arc<RwLock<UpstreamPool>>>,
-        routing_index: &RouteIndex,
-        header_lookup: Option<&LbHeaderLookup<'_>>,
-    ) -> Result<ResolvedBackend, ProxyError> {
-        if method.is_empty() || path.is_empty() {
-            return Err(ProxyError::Transport("empty method or path".into()));
-        }
-
-        let route_decision = routing_index
-            .lookup_with_decision_for_method(path, authority, Some(method))
-            .ok_or_else(|| ProxyError::Transport(format!("no route for {path}")))?;
-        let upstream_name = route_decision.upstream;
-
-        let upstream_pool = upstream_pools
-            .get(upstream_name)
-            .ok_or_else(|| ProxyError::Transport(format!("pool not found: {upstream_name}")))?
-            .clone();
-
-        let (backend_index, lb_type, backend_addr) = {
-            let (read_lb_type, read_fast_selected) = {
-                let pool = upstream_pool
-                    .read()
-                    .map_err(|_| ProxyError::Transport("upstream pool lock poisoned".into()))?;
-                if pool.pool.is_empty() {
-                    return Err(ProxyError::Transport("no servers in upstream".into()));
-                }
-                let lb_type = pool.lb_name();
-                let key = Self::resolve_lb_request_key(
-                    lb_type,
-                    pool.lb_key(),
-                    method,
-                    path,
-                    authority,
-                    cid_key,
-                    header_lookup,
-                );
-                // When a passively-ejected backend is pending re-admission, skip
-                // the read-only fast path so the write-locked slow path can
-                // reconcile it back into rotation.
-                let fast_selected = if pool.pool.readmit_due() {
-                    None
-                } else {
-                    pool.pick_readonly(key.as_str())
-                        .and_then(|idx| pool.pool.address(idx).map(|addr| (idx, addr.to_string())))
-                        .and_then(|(idx, addr)| {
-                            pool.begin_request_if_healthy(idx).then_some((idx, addr))
-                        })
-                };
-                (lb_type, fast_selected)
-            };
-
-            if let Some((idx, addr)) = read_fast_selected {
-                (idx, read_lb_type, addr)
-            } else {
-                let mut pool = upstream_pool
-                    .write()
-                    .map_err(|_| ProxyError::Transport("upstream pool lock poisoned".into()))?;
-                if pool.pool.is_empty() {
-                    return Err(ProxyError::Transport("no servers in upstream".into()));
-                }
-                let lb_type = pool.lb_name();
-                let key = Self::resolve_lb_request_key(
-                    lb_type,
-                    pool.lb_key(),
-                    method,
-                    path,
-                    authority,
-                    cid_key,
-                    header_lookup,
-                );
-
-                let idx = pool.pick(key.as_str()).ok_or_else(|| {
-                    let total = pool.pool.len();
-                    let healthy = pool.pool.healthy_len();
-                    error!(
-                        "no healthy backends available: {}/{} backends healthy",
-                        healthy, total
-                    );
-                    ProxyError::Transport("no healthy servers".into())
-                })?;
-                let backend_addr = pool
-                    .pool
-                    .address(idx)
-                    .map(str::to_string)
-                    .ok_or_else(|| ProxyError::Transport("invalid server address".into()))?;
-                (idx, lb_type, backend_addr)
-            }
-        };
-
-        debug!(
-            "Selected backend {} via {} route={} path_len={} host_specific={} reason={:?}",
-            backend_addr,
-            lb_type,
-            upstream_name,
-            route_decision.matched_path_len,
-            route_decision.host_specific,
-            route_decision.reason
-        );
-        Ok(ResolvedBackend {
-            upstream_name: upstream_name.to_string(),
-            backend_addr,
-            backend_index,
-            upstream_pool,
-            backend_lb: lb_type.to_string(),
-            route_path_len: route_decision.matched_path_len,
-            route_host_specific: route_decision.host_specific,
-            route_reason: route_decision.reason,
-        })
-    }
-
     fn resolve_lb_key_from_spec(
         lb_key_spec: &str,
         method: &str,
